% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSPmix.Rmd.
\name{aic_fspmix}
\alias{aic_fspmix}
\title{Using AIC to choose the best h_0}
\usage{
aic_fspmix(
  data,
  label,
  max_new_clust = 1,
  bandwidth = 1,
  max_iter = 1000,
  min_gap = 0.1,
  nrep = 10
)
}
\arguments{
\item{data}{Data in matrix form(n * p). Can be the data result from prepare_data}

\item{label}{Label set. Can be the label result from prepare_data}

\item{max_new_clust}{Maximum number of new cluster. Default is 1}

\item{bandwidth}{Smooth bandwidth}

\item{max_iter}{Maximum iteration}

\item{min_gap}{Minimum difference of likelihood}

\item{nrep}{Number of repetitions}
}
\description{
Using AIC to choose the best h_0
}