% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSPmix.Rmd.
\name{comparison_visual}
\alias{comparison_visual}
\title{Model results comparison}
\usage{
comparison_visual(res1, res2)
}
\arguments{
\item{res1}{Output of fspmix function}

\item{res2}{Output of fspmix function}
}
\description{
Model results comparison
}