% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSPmix.Rmd.
\name{create_hpb}
\alias{create_hpb}
\title{Draw high probability bands}
\usage{
create_hpb(res, alpha = 0.05, label = NULL)
}
\arguments{
\item{res}{Output from fspmix}

\item{alpha}{Alpha level, default is 0.05}

\item{label}{Label set, used to assign graph titles. If empty, titles will be automatically given Group 1 to K}
}
\description{
Draw high probability bands
}