% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-FSPmix.Rmd.
\name{visualize_res_UMAP}
\alias{visualize_res_UMAP}
\title{Prediction result in UMAP space}
\usage{
visualize_res_UMAP(data, label, res)
}
\arguments{
\item{data}{Data}

\item{label}{Label}

\item{prediction}{Generate from prediction function}
}
\description{
Prediction result in UMAP space
}